% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{isrcs}
\alias{isrcs}
\title{Test if \code{\link[rms]{rcs}} from the \code{rms} Package was Used in the Fit.}
\usage{
isrcs(fit)
}
\arguments{
\item{fit}{a fit object of class \code{lm}, \code{glm}, \code{coxph}, or \code{mfp}.}
}
\value{
A list with two elements:
\tabular{ll}{
    [[1]] \tab \code{\link[rms]{rcs}} is included in formula (right-hand side), but not as a variable name.\cr
    [[2]] \tab variables with \code{\link[rms]{rcs}}\cr
}
}
\description{
Test if \code{\link[rms]{rcs}} from the \code{rms} Package was Used in the Fit.
}
\keyword{internal}
