\name{plot.shrink}
\alias{plot.shrink}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
plot method for class 'shrink'
}
\description{
Plots an object of class \code{shrink}. Two types of plots are available
 which an be selected using the argumet \code{which}. \code{which = 1} 
 returns a Residual vs Fitted plot, and \code{which = 2} gives a normal 
 probability plot for the residuals.
}
\usage{
plot.shrink(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
An object of class \code{shrink}, usually obtained from \code{shrink()}
}
  \item{\dots}{
More argumets that can be passed to \code{plot()}
}
}
\details{
TBA
}
%%\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
%%}
%%\references{
%% ~put references to the literature/web site here ~
%%}
\author{
Enayetur Raheem
}
\note{
Note TBA
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{plot.lm}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
