\name{print.shrink}

\alias{print.shrink}

\title{ Print Method for Objects of Class \code{shrink} }

\description{ 
  This class of objects is returned by the \code{shrink} function. 
  Objects of this class have methods for the functions \code{print}, \code{predict},
  and \code{coeff}. 
}

\usage{
\method{print}{shrink}(x, ...) 
}

\arguments{
  \item{x}{ object of class \code{shrink}. }
  \item{...}{ further arguments. }  
}

%%\details{ }

%%\value{ }

%%\references{  }

\author{ Daniela Dunkler, Georg Heinze }

%%\note{  }

\seealso{ \code{\link{shrink}}, \code{\link{coef.shrink}} }

%%\examples{ }

\keyword{ utilities }
