% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.shrink.R
\name{coef.shrink}
\alias{coef.shrink}
\title{Returns Shrunken Regression Coefficients from Objects of Class \code{shrink}}
\usage{
\method{coef}{shrink}(object, ...)
}
\arguments{
\item{object}{an object of class \code{shrink}.}

\item{...}{further arguments.}
}
\value{
A vector with shrunken regression coefficients
}
\description{
This class of objects is returned by the \code{shrink} function. Objects of this
class have methods for the functions \code{coef}, \code{predict}, \code{print},
\code{summary}, and \code{vcov}.
}
\seealso{
\code{\link{shrink}}, \code{\link{print.shrink}}, \code{\link{predict.shrink}},
    \code{\link{summary.shrink}}, \code{\link{vcov.shrink}}
}

