% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.shrink.R
\name{print.shrink}
\alias{print.shrink}
\title{Print Method for Objects of Class \code{shrink}}
\usage{
\method{print}{shrink}(x, ...)
}
\arguments{
\item{x}{object of class \code{shrink}.}

\item{...}{further arguments.}
}
\description{
This class of objects is returned by the \code{shrink} function. Objects of this
class have methods for the functions \code{coef}, \code{predict}, \code{print},
\code{summary}, and \code{vcov}.
}
\seealso{
\code{\link{shrink}}, \code{\link{coef.shrink}}, \code{\link{predict.shrink}},
    \code{\link{summary.shrink}}, \code{\link{vcov.shrink}}
}
