% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.shrink.R
\name{summary.shrink}
\alias{summary.shrink}
\title{Summary Method for Objects of Class \code{shrink}}
\usage{
\method{summary}{shrink}(object, digits = 6, ...)
}
\arguments{
\item{object}{an object of class \code{shrink}.}

\item{digits}{integer, used for number formatting with \code{\link[base]{signif}}().}

\item{...}{further arguments.}
}
\value{
A matrix with regression coefficients of the orginial fit, corresponding
    shrinkage factors and shrunken regression coefficients.
}
\description{
This class of objects is returned by the \code{shrink} function. Objects of this
class have methods for the functions \code{coef}, \code{predict}, \code{print},
\code{summary}, and \code{vcov}.
}
\seealso{
\code{\link{shrink}}, \code{\link{coef.shrink}}, \code{\link{print.shrink}},
    \code{\link{predict.shrink}}, \code{\link{vcov.shrink}}
}

